refactor: moved implementation into other file
